Update database configuration
This commit updates the database configuration in the `.env.sample` file and the `database.js` file. The `DB_USER` value in the `.env.sample` file has been changed from `"root@127.0.0.1"` to `"root"`. In the `database.js` file, the `host`, `username`, `password`, and `database` values have been hardcoded to `'127.0.0.1'`, `"root"`, `''`, and `'esamyak'` respectively.